Clean up the codebase
Removed unused methods from Fn
Rearrange the TypesI and Container classes
Begin documentation
