+++ core/src/fr/hammons/slinc/Slinc.scala
@@ -23,10 +23,11 @@ trait Slinc:
   protected val layoutI = LayoutI(layoutPlatformSpecific)
   protected val transitionsI = TransitionsI(transitionsPlatformSpecific)
   protected val structI = StructI(layoutI, transitionsI, jitManager)
-  protected val typesI = TypesI.platformTypes(layoutI)
+  val typesI = types.TypesI.platformTypes(layoutI)
   protected val scopeI = ScopeI(scopePlatformSpecific)
   protected val libraryI = LibraryI(libraryIPlatformSpecific)
   val receiveI = ReceiveI(libraryIPlatformSpecific, layoutI)
+
   export layoutI.{*, given}
   export typesI.{*, given}
   export libraryI.*
@@ -35,7 +36,7 @@ trait Slinc:
   export transitionsI.given
   export structI.Struct
   export scopeI.given
-  export ContextProof.given
+  export container.ContextProof.given
   export receiveI.given
 
   def sizeOf[A](using l: LayoutOf[A]) = l.layout.size.toLong.maybeAs[SizeT]
