+++ core/src/fr/hammons/slinc/Variadic.scala
@@ -1,0 +1,6 @@
+package fr.hammons.slinc
+
+import container.*
+
+type Variadic = Container[LayoutOf *::: NativeInCompatible *::: End]
+
