+++ core/src/fr/hammons/slinc/container/Capabilities.scala
@@ -1,0 +1,5 @@
+package fr.hammons.slinc.container
+
+private[slinc] sealed trait Capabilities
+private[slinc] sealed trait *:::[A[_], B <: Capabilities] extends Capabilities
+private[slinc] sealed trait End extends Capabilities
