+++ core/src/fr/hammons/slinc/types/OS.scala
@@ -1,4 +1,4 @@
-package fr.hammons.slinc
+package fr.hammons.slinc.types
 
 private[slinc] enum OS:
   case Linux
